Move last modified into protobuf inode

+++ core/server/master/src/main/java/alluxio/master/file/DefaultFileSystemMaster.java
@@ -1440,7 +1440,7 @@ public final class DefaultFileSystemMaster extends CoreMaster implements FileSys
   }
 
   @Override
-  public int estimateNumberOfPaths() {
+  public long estimateNumberOfPaths() {
     return mInodeTree.estimateSize();
   }
