+++ core/server/master/src/main/java/alluxio/master/metastore/ReadOnlyInodeStore.java
@@ -23,7 +23,7 @@ public interface ReadOnlyInodeStore {
   /**
    * @return an estimate for the number of inodes in the inode store
    */
-  int estimateSize();
+  long estimateSize();
 
   /**
    * @param id an inode id
